[OPIK-7437] [FE] refactor: extract swipe/animation logic into useSwipeableSteps hook
MobileOnboardingShell is now pure markup: the scroll/step sync state
machine, entrance/reverse animation sequencing, restore timers and the
inert handling live in useSwipeableSteps, consumed via three prop
getters (scrollerProps, getPanelProps, getContentProps).
